ref(onboarding): Create php-laravel folder
