+++ static/app/gettingStartedDocs/php-laravel/onboarding.spec.tsx
@@ -4,7 +4,7 @@ import {textWithMarkupMatcher} from 'sentry-test/utils';
 
 import {ProductSolution} from 'sentry/components/onboarding/gettingStartedDoc/types';
 
-import docs from './laravel';
+import docs from './index';
 
 describe('laravel onboarding docs', () => {
   it('renders doc correctly', () => {
