+++ static/app/utils/gettingStartedDocs/getPlatformPath.ts
@@ -1,11 +1,22 @@
 import type {PlatformIntegration} from 'sentry/types/project';
 
+/**
+ * Platforms using new folder structure (platform/onboarding.tsx, platform/logs.tsx, etc.) instead of legacy (language/framework).
+ */
+const MIGRATED_PLATFORMS = new Set(['php-laravel']);
+
 export function getPlatformPath(platform: PlatformIntegration) {
   // handle console platforms (xbox, playstation, nintendo-switch, etc.)
   if (platform.type === 'console') {
     return `console/${platform.id}`;
   }
 
+  // Check if platform has been migrated to new structure
+  if (MIGRATED_PLATFORMS.has(platform.id)) {
+    return platform.id;
+  }
+
+  // Legacy structure for platforms not yet migrated
   // some platforms use a naming convention that combines 'language' and 'id' with a hyphen in between. For example, 'react-native'.
   if (platform.id === platform.language) {
     return `${platform.language}/${platform.language}`;
